from BigDecimal to BigInt bc of weis

+++ src/transaction.ts
@@ -27,10 +27,10 @@ export function newTransaction(event: Transfer): void {
                 event.params.from, 
                 event.params.to, 
                 event.address.toHexString(), 
-                event.params.value.toBigDecimal(), 
+                event.params.value, 
                 event.params.data, 
                 event.block.timestamp, 
-                event.transaction.gasUsed.toBigDecimal().times(event.transaction.gasPrice.toBigDecimal()),
+                event.transaction.gasUsed.times(event.transaction.gasPrice),
                 false
             );
         }
@@ -40,11 +40,11 @@ export function newTransaction(event: Transfer): void {
     }
 
     if (event.params.from == Address.fromI32(0)) {
-        handleTokenMint(event.address.toHexString(), event.params.value.toBigDecimal());
+        handleTokenMint(event.address.toHexString(), event.params.value);
     }
 
     if (event.params.to == Address.fromI32(0)) {
-        handleTokenBurn(event.address.toHexString(), event.params.value.toBigDecimal());
+        handleTokenBurn(event.address.toHexString(), event.params.value);
     }
 }
 
@@ -53,10 +53,10 @@ export function createTransaction(
     from: Address,
     to: Address,
     currency: string,
-    amount: BigDecimal,
+    amount: BigInt,
     data: Bytes,
     timestamp: BigInt,
-    fee: BigDecimal,
+    fee: BigInt,
     isBankTransaction: boolean
 ): 
     Transaction
